+++ src/nick/personal/constant/ConstantMethodTypeInfo.java
@@ -9,9 +9,9 @@ public class ConstantMethodTypeInfo extends ConstantInfo {
     /**
      * 只必须是对常量池的有效索引，常量池在该索引处的项必须是 {@link ConstantUtf8Info} 结构，表示方法的描述符
      */
-    public short descriptorIndex;
+    public int descriptorIndex;
 
-    public ConstantMethodTypeInfo(byte tag) {
+    public ConstantMethodTypeInfo(int tag) {
         super(tag);
     }
